Added a grid + undo mode to the map editor, improved the debug thing

+++ GetXYRatio.cpp
@@ -1,7 +1,7 @@
 #include"GetXYRatio.h"
 #define LECONVERSION (3.1418 / 180)
 
-void GetXYRatio(float * x, float * y, float Angle, int SPEED)
+void GetXYRatio(float * x, float * y, float Angle, float SPEED)
 {
 	if (Angle > 360) Angle -= 360;
